# Add sync.pool to contain our token map in conflictingTokensExist