--- a/ring/model.go
+++ b/ring/model.go
@@ -303,10 +303,10 @@ func normalizeIngestersMap(inputRing *Desc) {
 	}
 }
 
-var tokenMapPool = sync.Pool{New: func() interface{} { return make(map[uint32]bool) }}
+var tokenMapPool = sync.Pool{New: func() interface{} { return make(map[uint32]struct{}) }}
 
 func conflictingTokensExist(normalizedIngesters map[string]InstanceDesc) bool {
-	tokensMap := tokenMapPool.Get().(map[uint32]bool)
+	tokensMap := tokenMapPool.Get().(map[uint32]struct{})
 	defer func() {
 		for k := range tokensMap {
 			delete(tokensMap, k)
@@ -315,10 +315,10 @@ func conflictingTokensExist(normalizedIngesters map[string]InstanceDesc) bool {
 	}()
 	for _, ing := range normalizedIngesters {
 		for _, t := range ing.Tokens {
-			if tokensMap[t] {
+			if _, contains := tokensMap[t]; contains {
 				return true
 			}
-			tokensMap[t] = true
+			tokensMap[t] = struct{}{}
 		}
 	}
 	return false