<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

**What this PR does**: This PR _should_ use a memory pool instead of allocating a map for every time we call conflictingTokensExists to allow us to gc that memory easily and prevent long lived maps.

**Which issue(s) this PR fixes**:

Fixes #<issue number>

**Checklist**
- [X] Tests updated
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`