@treid314 2021-11-30
Looking at a pprof with a base of main and my branch looks like this showing a huge improvement in CPU time here.
<img width="1250" alt="image" src="https://user-images.githubusercontent.com/855447/144135066-bbec10b8-3645-45d9-b274-c79ce735ef2a.png">
As well as a pretty big improvement in memory between main and my branch here.
<img width="784" alt="image" src="https://user-images.githubusercontent.com/855447/144135209-a27485f6-325c-4336-9a8a-1f12309b9cdc.png">
These numbers are so large I'm unsure if my change here is correct.