@treid314 2021-11-30
So when we clear the map via deleting all of the keys we do see a pretty decent CPU perf improvement compared to main
<img width="1103" alt="image" src="https://user-images.githubusercontent.com/855447/144140031-10dcb6c6-9a3d-4383-9ecf-f133c23aa864.png">
And a pretty large memory improvement over main because we're immediately clearing this map
<img width="763" alt="image" src="https://user-images.githubusercontent.com/855447/144140275-df736a61-7fc5-4d18-9c06-d060f69666cb.png">