@treid314 2021-11-30
Using a map[uint32]struct{} saves us a byte per entry, which in our testing was about 7 meg which is free memory!
<img width="949" alt="image" src="https://user-images.githubusercontent.com/855447/144145183-58e8d6de-10cd-463d-b1a3-fc0cd812a8fd.png">