@pracucci 2021-12-01
I've run `BenchmarkMemberlistReceiveWithRingDesc` and compared preformances with `main` version:

```
name                              old time/op    new time/op    delta
MemberlistReceiveWithRingDesc-12    3.24ms ± 2%    3.04ms ± 1%   -6.05%  (p=0.012 n=3+3)

name                              old alloc/op   new alloc/op   delta
MemberlistReceiveWithRingDesc-12    1.01MB ± 0%    0.05MB ±15%  -94.87%  (p=0.000 n=3+3)

name                              old allocs/op  new allocs/op  delta
MemberlistReceiveWithRingDesc-12       669 ± 0%       681 ± 1%     ~     (p=0.132 n=3+3)
```